+++ src/api-key/api-key.controller.ts
@@ -27,7 +27,7 @@ import {
 import { GetUser } from '../user/decorators/user.decorator'
 import { User } from '../user/entities/user.entity'
 import { CreateApiKeyResponseDto } from './dto/create-api-key-response.dto'
-import { CreateApiKeyDto } from './dto/create-apiKey.dto'
+import { CreateApiKeyDto } from './dto/create-api-key.dto'
 import { ApiKey } from './entities/api-key.entity'
 
 @ApiTags('api-keys')
